SM-204 + Edit person state in the group with change GroupPersonLog

+++ src/app/pages/groups/group-administration/requests/requests.component.ts
@@ -9,6 +9,11 @@ import {PropertyConstant} from '../../../../data/local/property-constant';
 import {GroupPersonQuery} from '../../../../data/remote/rest-api/query/group-person-query';
 import {NgxVirtualScrollComponent} from '../../../../components/ngx-virtual-scroll/ngx-virtual-scroll/ngx-virtual-scroll.component';
 import {Direction} from '../../../../components/ngx-virtual-scroll/model/direction';
+import {NgxModalService} from '../../../../components/ngx-modal/service/ngx-modal.service';
+import {EditGroupPersonLogComponent} from '../../component/edit-group-person-log/edit-group-person-log.component';
+import {Group} from '../../../../data/remote/model/group/base/group';
+import {SplitButtonItem} from '../../../../components/ngx-split-button/bean/split-button-item';
+import {NgxModalRef} from '../../../../components/ngx-modal/bean/ngx-modal-ref';
 
 @Component({
   selector: 'app-requests',
@@ -21,10 +26,12 @@ export class RequestsComponent implements OnInit {
   public ngxVirtualScrollComponent: NgxVirtualScrollComponent;
 
   public groupPersonQuery: GroupPersonQuery;
+  private _group: Group;
 
   constructor(private  _participantRestApiService: ParticipantRestApiService,
               private _groupService: GroupService,
-              private _appHelper: AppHelper) {
+              private _appHelper: AppHelper,
+              private _ngxModalService: NgxModalService) {
     this.groupPersonQuery = new GroupQuery();
     this.groupPersonQuery.name = '';
     this.groupPersonQuery.from = 0;
@@ -34,29 +41,64 @@ export class RequestsComponent implements OnInit {
   }
 
   async ngOnInit() {
+    this._group = this._groupService.getGroup();
     await this.updateItems();
   }
 
   public async onAdd(groupPerson: GroupPerson) {
-    await this._participantRestApiService.putApprovePersonInGroup(
-      {
-        id: this._groupService.getGroup().id,
-        personId: groupPerson.person.id
-      });
-    this._appHelper.removeItem(this.ngxVirtualScrollComponent.items, groupPerson);
+    await this.showModal(groupPerson, async (modal: NgxModalRef, component: EditGroupPersonLogComponent) => {
+      return {
+        nameKey: 'save',
+        callback: async () => {
+          await this._appHelper.trySave(async () => {
+            await this._participantRestApiService.putApprovePersonInGroup(
+              {
+                id: this._groupService.getGroup().id,
+                personId: groupPerson.person.id
+              });
+            this._appHelper.removeItem(this.ngxVirtualScrollComponent.items, groupPerson);
+            await this._participantRestApiService.updateGroupPersonLog(component.data, {}, {groupId: this._group.id, personId: groupPerson.person.id, groupPersonLogId: component.data.id});
+            modal.dismiss();
+          });
+        }
+      };
+    });
   }
 
   public async onRemove(groupPerson: GroupPerson) {
-    await this._participantRestApiService.deleteApprovePersonInGroup({
-      id: this._groupService.getGroup().id,
-      personId: groupPerson.person.id
+    await this.showModal(groupPerson, async (modal: NgxModalRef, component: EditGroupPersonLogComponent) => {
+      return {
+        nameKey: 'remove',
+        callback: async () => {
+          await this._appHelper.tryRemove(async () => {
+            await this._participantRestApiService.updateGroupPersonLog(component.data, {}, {groupId: this._group.id, personId: groupPerson.person.id, groupPersonLogId: component.data.id});
+            await this._participantRestApiService.deleteApprovePersonInGroup({
+              id: this._groupService.getGroup().id,
+              personId: groupPerson.person.id
+            });
+            this._appHelper.removeItem(this.ngxVirtualScrollComponent.items, groupPerson);
+            modal.dismiss();
+          });
+        }
+      };
     });
-    this._appHelper.removeItem(this.ngxVirtualScrollComponent.items, groupPerson);
   }
 
   public getItems: Function = async (direction: Direction, pageQuery: PageQuery) => {
     return await this._participantRestApiService.getGroupPersonsByGroup(pageQuery);
   };
+
+  private async showModal(groupPerson: GroupPerson, initializeSplitButtonItem: (modal: NgxModalRef, component: EditGroupPersonLogComponent) => Promise<SplitButtonItem>) {
+    const modal = this._ngxModalService.open();
+    modal.componentInstance.titleKey = 'edit';
+
+    await modal.componentInstance.initializeBody(EditGroupPersonLogComponent, async component => {
+      component.manualInitialization = true;
+      const groupPersonLog = await this._participantRestApiService.getLatestGroupPersonLog({groupId: this._group.id, personId: groupPerson.person.id});
+      await component.initialize(groupPersonLog);
+      modal.componentInstance.splitButtonItems = [await initializeSplitButtonItem(modal, component)];
+    });
+  }
 
   private async updateItems() {
     setTimeout(async () => {
